Authentication completed

+++ backend/controllers/goalsController.js
@@ -1,13 +1,16 @@
 const asyncHandler = require ('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const goal = require('../models/userModel')
 
 
 //@desc  Getgoals
 //@route GET/api/goals
 //@access Private
 const getGoals = asyncHandler(async (req, res) => {
-const goals = await Goal.find()
+const goals = await Goal.find({
+    user:req.user.id
+})
 
     res.status(200).json(goals)
 })
@@ -22,6 +25,7 @@ const setGoal = asyncHandler(async(req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id,
     })
         
     res.status(200).json(goal)
@@ -39,6 +43,20 @@ const updateGoal = asyncHandler(async(req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+     const user = await User.findById(req.user.id)
+
+    //check for user
+     if(!user){
+        res.status(401)
+        throw new Error ('User not found')
+
+     }
+     // make sure the logged in user matches the goal user
+     if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error("User not authorised")
+     }
 
     const updatedGoal = await Goal. findByIdAndUpdate(req.params.id, req.body.text,{
         new: true,
@@ -58,6 +76,25 @@ const deleteGoal = asyncHandler(async(req, res) => {
         res.status(400)
         throw new Error("Goal not found")
     }
+    const user = await User.findById(req.user.id)
+
+    //check for user
+     if(!user){
+        res.status(401)
+        throw new Error ('User not found')
+
+     }
+     // make sure the logged in user matches the goal user
+     if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error("User not authorised")
+     }
+
+    const updatedGoal = await Goal. findByIdAndUpdate(req.params.id, req.body.text,{
+        new: true,
+    } )
+
+
     await Goal.findByIdAndRemove()
 
     res.status(200).json({id: req.params.id})
